+++ src/scripts/interfaces/choice-full.ts
@@ -19,4 +19,5 @@ export interface ChoiceFull {
   selected: boolean;
   value: string;
   score: number;
+  rank: number;
 }