+++ test/scripts/actions/choices.test.ts
@@ -13,6 +13,7 @@ describe('actions/choices', () => {
     id: 1,
     groupId: 1,
     score: 0,
+    rank: 0,
     disabled: false,
     elementId: '1',
     labelClass: stringToHtmlClass('test foo--bar'),