+++ test/scripts/search/index.test.ts
@@ -1,7 +1,9 @@
 import { expect } from 'chai';
+import { beforeEach } from 'vitest';
 import { searchByPrefixFilter } from '../../../src/scripts/search/prefix-filter';
 import { searchByFuse } from '../../../src/scripts/search/fuse';
 import { DEFAULT_CONFIG } from '../../../src';
+import { cloneObject } from '../../../src/scripts/lib/utils';
 
 export interface SearchableShape {
   label: string;
@@ -20,7 +22,9 @@ describe('search', () => {
 
   describe('fuse-full', () => {
     const search = searchByFuse;
-    process.env.SEARCH_FUSE = 'full';
+    beforeEach(() => {
+      process.env.SEARCH_FUSE = 'full';
+    });
     it('empty result', () => {
       const results = search<SearchableShape>(options, haystack, '');
       expect(results.length).eq(0);
@@ -36,12 +40,40 @@ describe('search', () => {
         `${haystack.length - 1}`,
       );
       expect(results.length).eq(1);
+    });
+
+    it('search order', () => {
+      const results = search<SearchableShape>(options, haystack, 'label');
+
+      expect(results.length).eq(haystack.length);
+      haystack.forEach((value, index) => {
+        expect(results[index].item.value).eq(value.value);
+      });
+    });
+
+    it('search order - custom sortFn', () => {
+      const opts = cloneObject(options);
+      opts.fuseOptions.sortFn = (a, b) => {
+        if (a.score === b.score) {
+          return a.idx < b.idx ? 1 : -1;
+        }
+
+        return a.score < b.score ? 1 : -1;
+      };
+      const results = search<SearchableShape>(opts, haystack, 'label');
+
+      expect(results.length).eq(haystack.length);
+      haystack.reverse().forEach((value, index) => {
+        expect(results[index].item.value).eq(value.value);
+      });
     });
   });
 
   describe('fuse-basic', () => {
     const search = searchByFuse;
-    process.env.SEARCH_FUSE = 'basic';
+    beforeEach(() => {
+      process.env.SEARCH_FUSE = 'basic';
+    });
     it('empty result', () => {
       const results = search<SearchableShape>(options, haystack, '');
       expect(results.length).eq(0);
@@ -57,6 +89,14 @@ describe('search', () => {
         `${haystack.length - 1}`,
       );
       expect(results.length).eq(1);
+    });
+    it('search order', () => {
+      const results = search<SearchableShape>(options, haystack, 'label');
+
+      expect(results.length).eq(haystack.length);
+      haystack.forEach((value, index) => {
+        expect(results[index].item.value).eq(value.value);
+      });
     });
   });